[Fix] Token, Badge, User 키 칼럼 변경
LONG(BIGINT)를 INT로 바꿨습니다.

+++ backend/UserServer/src/main/java/com/kl1verse/UserServer/domain/badge/repository/BadgeRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BadgeRepository extends JpaRepository<Badge, Integer> {
-    List<Badge> findByUserId(Long userId);
+    List<Badge> findByUserId(Integer userId);
 }
